Changed the UI for notes

+++ MusicJournal/ViewControllers/MyRecordingsTableViewController.swift
@@ -21,15 +21,6 @@ class MyRecordingsTableViewController: UITableViewController{
     
     @IBOutlet var myTableView: UITableView!
     
-    class Movie {
-        let name: String
-        var date: Int?
-        
-        init(_ name: String) {
-            self.name = name
-        }
-    }
-    
     static var chosenNumber: Int!
     
     @IBOutlet weak var songButton: UIButton!
@@ -131,7 +122,12 @@ class MyRecordingsTableViewController: UITableViewController{
             cell.lastModified.text="No Date"
         }
         
-        cell.pressPlayFile = currentRecording.filename!
+        if let thing = currentRecording.filename{
+            cell.pressPlayFile = currentRecording.filename!
+        } else{
+            cell.pressPlayFile = currentRecording.filename
+        }
+        
         
         let redColor = UIColor(red: 232/255, green: 90/255, blue: 69/255, alpha: 1)
         let lightBeigeBackground = UIColor(red: 234/255, green: 231/255, blue: 220/255, alpha: 1)
@@ -174,42 +170,43 @@ class MyRecordingsTableViewController: UITableViewController{
     
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath){
         if editingStyle == .delete{
-            
-            // Got the following code from: swiftdeveloperblog.com/code-examples/delete-file-example-in-swift/
-            // Find documents directory on device
-            let fileNameToDelete = ("\(arrayOfRecordingsInfo[indexPath.row].filename!)")
-            var filePath = ""
-            
-            let dirs : [String] = NSSearchPathForDirectoriesInDomains(FileManager.SearchPathDirectory.documentDirectory, FileManager.SearchPathDomainMask.allDomainsMask, true)
-            
-            if dirs.count > 0 {
-                let dir = dirs[0] //documents directory
-                filePath = dir.appendingFormat("/" + fileNameToDelete)
-                print("Local path = \(filePath)")
+            if arrayOfRecordingsInfo[indexPath.row].filename != nil{
+                // Got the following code from: swiftdeveloperblog.com/code-examples/delete-file-example-in-swift/
+                // Find documents directory on device
                 
-            } else {
-                print("Could not find local directory to store file")
-                return
-            }
-            
-            
-            do {
-                let fileManager = FileManager.default
+                let fileNameToDelete = ("\(arrayOfRecordingsInfo[indexPath.row].filename!)")
+                var filePath = ""
                 
-                // Check if file exists
-                if fileManager.fileExists(atPath: filePath) {
-                    // Delete file
-                    try fileManager.removeItem(atPath: filePath)
+                let dirs : [String] = NSSearchPathForDirectoriesInDomains(FileManager.SearchPathDirectory.documentDirectory, FileManager.SearchPathDomainMask.allDomainsMask, true)
+                
+                if dirs.count > 0 {
+                    let dir = dirs[0] //documents directory
+                    filePath = dir.appendingFormat("/" + fileNameToDelete)
+                    print("Local path = \(filePath)")
+                    
                 } else {
-                    print("for deleting, File does not exist")
+                    print("Could not find local directory to store file")
+                    return
                 }
                 
+                
+                do {
+                    let fileManager = FileManager.default
+                    
+                    // Check if file exists
+                    if fileManager.fileExists(atPath: filePath) {
+                        // Delete file
+                        try fileManager.removeItem(atPath: filePath)
+                    } else {
+                        print("for deleting, File does not exist")
+                    }
+                    
+                }
+                catch let error as NSError {
+                    print("An error took place: \(error)")
+                }
+                // End of code for deleting from the document directory also
             }
-            catch let error as NSError {
-                print("An error took place: \(error)")
-            }
-            // End of code for deleting from the document directory also
-            
             let recordingToDelete=arrayOfRecordingsInfo[indexPath.row]
             CoreDataHelper.deleteRecording(recording: recordingToDelete)
             arrayOfRecordingsInfo=CoreDataHelper.retrieveRecording()
